[APP] fix reconcilation error modal after rebase
